backend/requirements.txt
# Web Framework
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.15
psycopg2-binary==2.9.6

# MQTT / Messaging
paho-mqtt==1.6.1

# Data Processing
numpy==1.24.3

# Testing
pytest==7.3.1
httpx==0.24.1